requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
django-cors-headers==3.13.0
django-filter==22.1
django==4.0.6; python_version >= '3.8'
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
mysqlclient==2.1.1
pillow==9.2.0
pyjwt==2.4.0; python_version >= '3.6'
pytz==2022.1
sqlparse==0.4.2; python_version >= '3.5'
tzdata==2022.1; sys_platform == 'win32'
